// packages/ui/src/components/atoms/Button/Button.types.ts
export type EButtonTypes = "button" | "submit";

export type TButtonColors =
    | "transparent"
    | "transparentInverted"
    | "accentPrimary"
    | "accentSecondary"
    | "white"
    | "blue"
    | "green"
    | "red"
    | "grayLight"
    | "grayDark";

export type TButtonSizes = "small" | "medium" | "big";